Text Generator/requirements.txt
streamlit
google-generativeai
os
python-dotenv
